refactor: split out large models
